Update gui.py
made changes to the Button(.. , .. , .. , command = **** ). Changed the functions.. Instead of directly calling the functions in ie, we created methods in the class myGUI which then call the corresponding methods..
All the labels & buttons work properly.. Now we've to include Entry() in the GUI to replace taking input from the terminal.